// presantation_maker/src/types/const_for_presantation.d.ts
import { Background, TextDecoration, TextAlign } from "./types_of_presantation";
export declare const TEXT_OBJECT_DEFAULTS: {
    WIDTH: number;
    HEIGHT: number;
    FONT_SIZE: number;
    FONT_FAMILY: string;
    FONT_WEIGHT: number;
    TEXT_DECORATION: TextDecoration;
    TEXT_ALIGN: TextAlign;
    COLOR: string;
    SHADOW: {
        COLOR: string;
        BLUR: number;
    };
};
export declare const IMAGE_OBJECT_DEFAULTS: {
    WIDTH: number;
    HEIGHT: number;
};
export declare const DEFAULT_BACKGROUND: Background;
export declare const DEFAULT_POSITIONS: {
    X: number;
    Y: number;
};
//# sourceMappingURL=const_for_presantation.d.ts.map